Code refactoring andd also fixed issues for private key usage in sftp
