+++ app/src/main/kotlin/org/fossify/filemanager/enums/Authentication.kt
@@ -2,5 +2,7 @@ package org.fossify.filemanager.enums
 
 enum class Authentication {
     Password,
-    Anonymous
+    Anonymous,
+    PrivateKey
 }
+
