+++ app/src/main/kotlin/org/fossify/filemanager/interfaces/SMBApi.kt
@@ -1,0 +1,12 @@
+package org.fossify.filemanager.interfaces
+
+import jcifs.smb.SmbFile
+import org.fossify.filemanager.models.NetworkConnection
+
+interface SMBApi {
+    suspend fun verifyConnection(connection: NetworkConnection): Boolean
+
+    fun getFilesFromNetworkPath(): Array<SmbFile>
+
+    fun getMainSmbFile(): SmbFile
+}
