+++ app/src/main/res/layout/dialog_add_connection.xml
@@ -1,9 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:id="@+id/add_connection_dialog"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
-
+    <ScrollView
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:fillViewport="true">
     <LinearLayout
         android:id="@+id/change_view_type_dialog_holder"
         android:layout_width="match_parent"
@@ -123,6 +127,36 @@
                 </com.google.android.material.textfield.TextInputLayout>
 
                 <com.google.android.material.textfield.TextInputLayout
+                    android:id="@+id/privateKey_tf"
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    app:endIconMode="custom"
+                    android:visibility="gone"
+                    app:endIconDrawable="@drawable/baseline_insert_drive_file_24"
+                    android:hint="Private Key">
+
+                    <com.google.android.material.textfield.TextInputEditText
+                        android:id="@+id/privateKey_et"
+                        android:layout_width="match_parent"
+                        android:layout_height="wrap_content" />
+
+                </com.google.android.material.textfield.TextInputLayout>
+
+                <com.google.android.material.textfield.TextInputLayout
+                    android:id="@+id/privateKeyPass_tf"
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    android:visibility="gone"
+                    android:hint="Password Private Key">
+
+                    <com.google.android.material.textfield.TextInputEditText
+                        android:id="@+id/privateKeyPass_et"
+                        android:layout_width="match_parent"
+                        android:layout_height="wrap_content" />
+
+                </com.google.android.material.textfield.TextInputLayout>
+
+                <com.google.android.material.textfield.TextInputLayout
                     android:id="@+id/display_tf"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
@@ -195,4 +229,5 @@
         </LinearLayout>
 
     </LinearLayout>
+    </ScrollView>
 </androidx.constraintlayout.widget.ConstraintLayout>
